Move FindCardCompanyController out from Api namespace
The original idea was to have this validator behave as an
api returning a JSON response. This now behave like a normal
controller that renders views
