+++ config/routes.rb
@@ -3,9 +3,5 @@
 Rails.application.routes.draw do
   root 'home#index'
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
-  namespace :api do
-    namespace :v1 do
-      resources :find_card_company, only: :index
-    end
-  end
+  resources :find_card_company, only: :index
 end
