+++ spec/requests/validate_card_number_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe 'Find Card Company for Card Number' do
     let(:card_number) { '5485102976195757' }
 
     it 'returns a 200 response with the name of the card company' do
-      get '/api/v1/find_card_company', params: params
+      get '/find_card_company', params: params
 
       expect(response.body)
         .to include('Number belongs to Mastercard')
@@ -20,7 +20,7 @@ RSpec.describe 'Find Card Company for Card Number' do
     let(:card_number) { '5495102976195757' }
 
     it 'returns a 404 response with an error' do
-      get '/api/v1/find_card_company', params: params
+      get '/find_card_company', params: params
 
       expect(response.body)
         .to include('Invalid card number for Mastercard')
@@ -31,7 +31,7 @@ RSpec.describe 'Find Card Company for Card Number' do
     let(:card_number) { '1273519825814617' }
 
     it 'returns an error' do
-      get '/api/v1/find_card_company', params: params
+      get '/find_card_company', params: params
 
       expect(response.body)
         .to include('Unknown card number')
@@ -42,7 +42,7 @@ RSpec.describe 'Find Card Company for Card Number' do
     let(:card_number) { 'Not-a-number' }
 
     it 'returns an error' do
-      get '/api/v1/find_card_company', params: params
+      get '/find_card_company', params: params
 
       expect(response.body)
         .to include('Try inputting a number instead')
